Fix the unit test
